Add PluginMutex and move related types to its file. Updated MutexPlugin.
